backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
motor
python-dotenv
